update & refactoring: checkDataType 모듈화 및 refactoring 진행
